[ios-simulator Debug WK2] LayoutTest media/modern-media-controls/fullscreen-support/fullscreen-support-enabled.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=168668 <rdar://problem/30823677>
Patch by Antoine Quint <graouts@apple.com> on 2017-03-02
Reviewed by Simon Fraser.

This test would sometime fail for the first assertion because we may need more than just metadata
to determine that we can be in a state where the fullscreen button becomes enabled (see the logic
in FullscreenSupport). So we now use an asynchronous assertion to determine we've enabled the
fullscreen button and then force it to error to disable it.

* media/modern-media-controls/fullscreen-support/fullscreen-support-enabled-expected.txt:
* media/modern-media-controls/fullscreen-support/fullscreen-support-enabled.html:
* platform/ios-simulator/TestExpectations:

Identifier: 186132@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213348 268f45cc-cd09-0410-ab3c-d52691b4dbfc