Add TextIndicator support for providing a snapshot excluding selected content https://bugs.webkit.org/show_bug.cgi?id=169309 <rdar://problem/30883525>
Reviewed by Tim Horton.

Source/WebCore:

Work towards <rdar://problem/30882974>. Adds support in TextIndicator to include a snapshot of all visible
content, excluding nodes captured in the current selection. This comes in the form of the new
TextIndicatorOptionIncludeSnapshotOfAllVisibleContentWithoutSelection flag -- see per-method changes for more details.

No new tests, since there is no change in behavior yet.

* page/FrameSnapshotting.cpp:
(WebCore::snapshotFrameRectWithClip):
* page/FrameSnapshotting.h:

Introduce a new snapshot option flag, SnapshotOptionsPaintEverythingExcludingSelection.

(WebCore::snapshotFrameRectWithClip):
* page/TextIndicator.cpp:

If TextIndicatorOptionIncludeSnapshotOfAllVisibleContentWithoutSelection is turned on, capture and include a
snapshot of the visible bounds of the frame.

(WebCore::takeSnapshot):
(WebCore::takeSnapshots):
(WebCore::initializeIndicator):
* page/TextIndicator.h:

Introduce TextIndicatorOptionIncludeSnapshotOfAllVisibleContentWithoutSelection.

* platform/graphics/FontCascade.cpp:
(WebCore::FontCascade::widthOfTextRange):
* platform/graphics/FontCascade.h:

Teach FontCascade to retrieve the width of a range in a TextRun, and also compute the widths of the TextRun
before and after the range. This is used by InlineTextBox to compute the painting rect to clip out when skipping
text decorations for the selected region of the TextRun.

* rendering/InlineTextBox.cpp:

Teach InlineTextBox to respect PaintBehaviorExcludeSelection. When painting text, if this paint behavior is
used, we will not paint the portion of the text run that lies inside the range determined by selectionStartEnd().
Similarly, when painting text decorations, we exclude the rect surrounding the selected content.

(WebCore::InlineTextBox::paint):
(WebCore::InlineTextBox::paintDecoration):
* rendering/InlineTextBox.h:
* rendering/PaintPhase.h:

Introduce PaintBehaviorExcludeSelection, used to inform members of the render tree not to paint portions that
lie within the current selection range.

* rendering/RenderLayer.cpp:

Propagate the PaintBehaviorExcludeSelection flag to children when painting.

(WebCore::RenderLayer::paintLayerContents):
(WebCore::RenderLayer::paintForegroundForFragments):
* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::shouldPaint):

Teach RenderReplaced to respect PaintBehaviorExcludeSelection by bailing from painting if it is selected and the
paint behavior is enabled.

* rendering/TextDecorationPainter.h:
* rendering/TextPaintStyle.cpp:
(WebCore::computeTextSelectionPaintStyle):
* rendering/TextPaintStyle.h:
* rendering/TextPainter.cpp:
(WebCore::TextPainter::paintText):
* rendering/TextPainter.h:

Source/WebKit2:

Serialize the snapshot excluding the selection and the snapshot's bounds when sending TextIndicatorData over
IPC to the UI process.

* Shared/WebCoreArgumentCoders.cpp:
(IPC::ArgumentCoder<TextIndicatorData>::encode):
(IPC::ArgumentCoder<TextIndicatorData>::decode):


Identifier: 186371@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213614 268f45cc-cd09-0410-ab3c-d52691b4dbfc