[WebRTC] SDP sess-id in the "o=" line should be a value between 0 and LLONG_MAX. https://bugs.webkit.org/show_bug.cgi?id=169681
Reviewed by Alejandro G. Castro.

Source/WebCore:

Use an int64_t type for storing sess-id and ensure that the value generated is always between 0 and LLONG_MAX (2^63-1).
To the JS world (sdp.js) we pass and read this value as a string type.

The JS world (sdp.js) was already using a string type for sess-id.
This caused a bug when reading the sess-id value at configurationFromJSON(): no value was obtained because a numeric type was expected.

So, after this patch, sess-id is passed and read as a string from sdp.js, and converted to an int64_t type for internal use.
This way we also avoid a rounding problem that happens with values near LLONG_MAX when converting from int64_t to double when passed to the JS world,
and that could cause that we end passing a number bigger than LLONG_MAX to sdp.js.

* Modules/mediastream/SDPProcessor.cpp:
(WebCore::configurationFromJSON):
(WebCore::configurationToJSON):
* platform/mediastream/MediaEndpointSessionConfiguration.h:
(WebCore::MediaEndpointSessionConfiguration::sessionId):
(WebCore::MediaEndpointSessionConfiguration::setSessionId):
(WebCore::MediaEndpointSessionConfiguration::MediaEndpointSessionConfiguration):

LayoutTests:

Check that the generated sess-id value its between the limits before validating it.
This is covered by the tests fast/mediastream/RTCPeerConnection-inspect-answer.html
and fast/mediastream/RTCPeerConnection-inspect-offer.html.
Both tests will fail if the sess-id value generated is not within the limits.

* fast/mediastream/resources/sdp-utils.js:
(printComparableSessionDescription):

Identifier: 186709@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214037 268f45cc-cd09-0410-ab3c-d52691b4dbfc