Follow-up (r213833): write a layout test for 169199 https://bugs.webkit.org/show_bug.cgi?id=169576 <rdar://problem/31024766>
Reviewed by Said Abou-Hallawa.

Source/WebCore:

Add a function to Internals to force the deletion of the ImageDecoder just after
requesting the async decoding of a frame, and implement that behaviour in BitmapImage.

Test: fast/images/clear-animation-decoder.html

* platform/graphics/BitmapImage.cpp:
(WebCore::BitmapImage::internalStartAnimation):
* platform/graphics/BitmapImage.h:
* testing/Internals.cpp:
(WebCore::Internals::setClearDecoderAfterAsyncFrameRequestForTesting):
* testing/Internals.h:
* testing/Internals.idl:

LayoutTests:

Add a test that ensures there's no crash when deleting the ImageDecoder after requesting
an async decoding operation.

* fast/images/clear-animation-decoder-expected.txt: Added.
* fast/images/clear-animation-decoder.html: Added.


Identifier: 186764@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@214103 268f45cc-cd09-0410-ab3c-d52691b4dbfc